Update @types/jquery
Fixing the same problem as described in https://github.com/phovea/phovea_clue/issues/189 and fixed in https://github.com/phovea/phovea_clue/pull/190.

If tourdino and phovea_clue are used together in the same workspace the intersected version would be `3.3.33` which throws an error with the additional typings.

Switching to  `~3.3.33` results also in `~3.3.33` for the intersected version in the workspace package.json.